fix(create-webiny-project): improve CWP code structure and logging
